# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "RoundApp OpenApi Documentation",
        "description": "L5 Swagger OpenApi",
        "contact": {
            "email": "francesco.simeoli@gmail.com"
        },
        "license": {
            "name": "Apache 2.0 License",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://roundapp.local/api",
            "description": "RoundApp API Server"
        },
        {
            "url": "https://roundapp.local"
        }
    ],
    "paths": {
...
        "/campaigns": {
            "get": {
                "tags": [
                    "Campaigns"
                ],
                "summary": "Get Campaigns list",
                "description": "Get Campaigns list",
                "operationId": "indexCampaigns",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/CampaignResource"
                                            }
                                        },
                                        "pool": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "tags": {
                                                        "type": "array",
                                                        "items": {
                                                            "type": "object",
                                                            "additionalProperties": {
                                                                "type": "string"
                                                            }
                                                        }
                                                    },
                                                    "producers": {
                                                        "type": "array",
                                                        "items": {
                                                            "type": "object",
                                                            "additionalProperties": {
                                                                "type": "string"
                                                            }
                                                        }
                                                    },
                                                    "ages": {
                                                        "type": "array",
                                                        "items": {
                                                            "type": "object",
                                                            "additionalProperties": {
                                                                "type": "string"
                                                            }
                                                        }
                                                    },
                                                    "genders": {
                                                        "type": "array",
                                                        "items": {
                                                            "type": "object",
                                                            "additionalProperties": {
                                                                "type": "string"
                                                            }
                                                        }
                                                    },
                                                    "types": {
                                                        "type": "array",
                                                        "items": {
                                                            "type": "object",
                                                            "additionalProperties": {
                                                                "type": "string"
                                                            }
                                                        }
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Campaigns"
                ],
                "summary": "Create Campaign",
                "description": "Create Campaign",
                "operationId": "storeCampaigns",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#components/schemas/CampaignRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CampaignResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/campaigns/{id}": {
            "get": {
                "tags": [
                    "Campaigns"
                ],
                "summary": "Get Campaign",
                "description": "Get Campaign",
                "operationId": "showCampaign",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Campaign id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CampaignResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Campaigns"
                ],
                "summary": "Update Campaign",
                "description": "Update Campaign",
                "operationId": "updateCampaigns",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Campaign id",
                        "required": true
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#components/schemas/CampaignRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CampaignResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Campaigns"
                ],
                "summary": "Delete Campaign",
                "description": "Delete Campaign",
                "operationId": "deleteCampaigns",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Campaign id",
                        "required": true
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {}
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/campaigns/counters": {
            "get": {
                "tags": [
                    "Campaigns"
                ],
                "summary": "Get Campaigns counters",
                "description": "Get Campaigns counters",
                "operationId": "campaignsCounters",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/CampaignsCountersResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/campaigns/info": {
            "get": {
                "tags": [
                    "Campaigns"
                ],
                "summary": "Get Campaigns info",
                "description": "Get Campaigns info",
                "operationId": "infoCampaigns",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/CampaignInfoResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CampaignResource": {
                "description": "Resource Campaign",
                "required": [],
                "properties": {
                    "id": {
                        "schema": "CampaignResource",
                        "description": "id",
                        "type": "integer",
                        "format": "bigint"
                    },
                    "title": {
                        "schema": "CampaignResource",
                        "description": "title",
                        "type": "string",
                        "format": "string",
                        "maxLength": 100
                    },
                    "description": {
                        "schema": "CampaignResource",
                        "description": "description",
                        "type": "string",
                        "format": "text"
                    },
                    "tags": {
                        "schema": "CampaignResource",
                        "description": "tags",
                        "type": "array",
                        "format": "array",
                        "items": {
                            "$ref": "#/components/schemas/Tag"
                        }
                    },
                    "age": {
                        "schema": "CampaignResource",
                        "description": "age",
                        "type": "integer",
                        "format": "integer"
                    },
                    "gender": {
                        "schema": "CampaignResource",
                        "description": "gender",
                        "type": "string",
                        "enum": [
                            "male",
                            "female"
                        ]
                    },
                    "area": {
                        "schema": "CampaignResource",
                        "description": "area",
                        "type": "string"
                    },
                    "paymentType": {
                        "schema": "CampaignResource",
                        "description": "paymentType",
                        "type": "string",
                        "enum": [
                            "one",
                            "all",
                            "one_a_day"
                        ]
                    },
                    "people": {
                        "schema": "CampaignResource",
                        "description": "people",
                        "type": "integer"
                    },
                    "producer": {
                        "schema": "CampaignResource",
                        "$ref": "#/components/schemas/ProducerResource"
                    }
                },
                "type": "object"
            },
...
            "CampaignsCountersResource": {
                "description": "Resource CampaignsCountersResource",
                "properties": {
                    "videos": {
                        "schema": "CampaignsCountersResource",
                        "description": "Class CampaignsCountersResource",
                        "type": "integer",
                        "format": "bigint"
                    },
                    "producers": {
                        "schema": "CampaignsCountersResource",
                        "description": "Class CampaignsCountersResource",
                        "type": "integer",
                        "format": "bigint"
                    },
                    "costs": {
                        "schema": "CampaignsCountersResource",
                        "description": "Class CampaignsCountersResource",
                        "type": "integer",
                        "format": "bigint"
                    },
                    "views_global": {
                        "schema": "CampaignsCountersResource",
                        "description": "Class CampaignsCountersResource",
                        "type": "integer",
                        "format": "bigint"
                    },
                    "views_paid": {
                        "schema": "CampaignsCountersResource",
                        "description": "Class CampaignsCountersResource",
                        "type": "integer",
                        "format": "bigint"
                    },
                    "views_free": {
                        "schema": "CampaignsCountersResource",
                        "description": "Class CampaignsCountersResource",
                        "type": "integer",
                        "format": "bigint"
                    },
                    "clicks": {
                        "schema": "CampaignsCountersResource",
                        "description": "Class CampaignsCountersResource",
                        "type": "integer",
                        "format": "bigint"
                    }
                },
                "type": "object"
            },
...
            "Tag": {
                "description": "Entity Tag",
                "required": [
                    "id",
                    "code"
                ],
                "properties": {
                    "id": {
                        "schema": "Tag",
                        "description": "id",
                        "type": "integer",
                        "format": "bigint"
                    },
                    "code": {
                        "schema": "Tag",
                        "description": "code",
                        "type": "string",
                        "format": "string",
                        "maxLength": 200
                    },
                    "description": {
                        "schema": "Tag",
                        "description": "description",
                        "type": "string",
                        "format": "string",
                        "maxLength": 200
                    },
                    "created_at": {
                        "schema": "Tag",
                        "description": "created_at",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "schema": "Tag",
                        "description": "updated_at",
                        "type": "string",
                        "format": "date-time"
                    },
                    "deleted_at": {
                        "schema": "Tag",
                        "description": "deleted_at",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "passport": {
                "type": "oauth2",
                "description": "Laravel passport oauth2 security.",
                "in": "header",
                "scheme": "https",
                "flows": {
                    "password": {
                        "authorizationUrl": "https://roundapp.local/oauth/authorize",
                        "tokenUrl": "https://roundapp.local/oauth/token",
                        "refreshUrl": "https://roundapp.local/token/refresh",
                        "scopes": [],
                        "client_id": "1",
                        "secret": "yo4kJNkiGGIiU0YqZ5jP1ucCArLM11FC7z5QdEFY"
                    }
                }
            }
        }
    },
...
}